requirements.txt
click==7.0
PyGithub==1.43.3